tts_common/requirements.txt
# TTS Common Library Requirements

# TTS синтез
edge-tts>=6.1.0

# Работа с документами
python-docx>=1.0.0
pdfplumber>=0.10.0
striprtf>=0.0.26
EbookLib>=0.18

# Парсинг веб-страниц
trafilatura>=1.6.0

# HTML парсинг
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Асинхронность
aiofiles>=23.0.0
